requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# HTTP Client for microservices
httpx==0.25.2

# CyborgDB
cyborgdb==0.14.0



# Kafka Streaming (Phase 2: Secure Data Pipeline)
kafka-python==2.0.2

# Encryption
cryptography==41.0.7

# Scientific Computing
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3

# Advanced ML Models (Phase 3: Advanced Fraud Detection)
xgboost==2.0.2
tensorflow==2.15.0
joblib==1.3.2

# Production Optimization (Phase 4: Production Deployment)
onnxruntime==1.20.0
onnx==1.14.1
skl2onnx==1.15.0
prometheus-client==0.19.0
psutil==5.9.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Encryption (for production use)
cryptography==41.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0

# Monitoring
python-json-logger==2.0.7
